Introduce MakeReadOnlyOrImmutableObj
We use MakeReadOnlyOrImmutableObj, because MakeReadOnlyObj and
MakeImmutable don't play nicely together in HPC-GAP:
Applying MakeReadOnlyObj recursively makes an object `obj` and all its
subobjects read-only *as long as the subobjects* live in the same region
as `obj`.
Now if a compressed matrix is made immutable, then its rows are moved
into the public region. The compressed matrix though is a positional
object and as such is *not* moved into the public region.
Consequently, applying the MakeReadOnlyObj to a compressed matrix will
not make its rows read-only.